# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API EBAC",
        "description": "Documentation OpenAPI des API EBAC : églises, paramètres académiques et pré-inscription.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Serveur courant"
        },
        {
            "url": "https://api-ebac.severinzran.ci/api/v1",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/parametres/promotions": {
            "get": {
                "tags": [
                    "Promotions"
                ],
                "operationId": "listerPromotions",
                "parameters": [
                    {
                        "name": "niveau",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "annee",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "promotion",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste filtrée des promotions"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Promotions"
                ],
                "operationId": "creerPromotion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PromotionPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Promotion créée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/parametres/promotions/{id}": {
            "get": {
                "tags": [
                    "Promotions"
                ],
                "operationId": "afficherPromotion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Promotion trouvée"
                    },
                    "404": {
                        "description": "Introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Promotions"
                ],
                "operationId": "remplacerPromotion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PromotionPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Promotion modifiée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Promotions"
                ],
                "operationId": "supprimerPromotion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Promotion supprimée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Promotions"
                ],
                "operationId": "modifierPromotion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PromotionPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Promotion modifiée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "PromotionPayload": {
                "required": [
                    "code",
                    "id_annee_academique",
                    "id_niveau"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "PROMO-2026-A1"
                    },
                    "id_annee_academique": {
                        "type": "integer"
                    },
                    "id_niveau": {
                        "type": "integer"
                    },
                    "capacite": {
                        "type": "integer",
                        "nullable": true
                    },
                    "statut": {
                        "type": "string",
                        "default": "Active"
                    },
                    "date_ouverture": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "date_cloture": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Jeton retourné après la connexion et la confirmation OTP.",
                "bearerFormat": "Token Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}